requirements-gpu.txt
requests
keras>=2.1.4
tensorflow-gpu>=1.4.0